Rename `fmt.rs` to `format.rs`
