+++ src/lib.rs
@@ -17,8 +17,8 @@ pub mod fetchers;
 mod target;
 pub use target::*;
 
-mod fmt;
-pub use fmt::*;
+mod format;
+pub use format::*;
 
 /// Default package path template (may be overridden in package Cargo.toml)
 pub const DEFAULT_PKG_URL: &str =
